+++ pm4py/algo/discovery/inductive/variants/im/algorithm.py
@@ -31,6 +31,7 @@ from pm4py.objects.log.util import filtering_utils
 from pm4py.util import constants, xes_constants
 from pm4py.objects.log.log import EventLog, Trace, Event
 from pm4py.algo.discovery.inductive.util import tree_consistency
+from pm4py.util import variants_util
 import pkgutil
 
 
@@ -183,7 +184,7 @@ def apply_tree_variants(variants, parameters=None):
     var_keys = list(variants.keys())
     for var in var_keys:
         trace = Trace()
-        activities = var.split(constants.DEFAULT_VARIANT_SEP)
+        activities = variants_util.get_activities_from_variant(var)
         for act in activities:
             trace.append(Event({activity_key: act}))
         log.append(trace)